Add readComponent
